+++ src/analyse.py
@@ -3,7 +3,7 @@ import logging
 import pandas as pd
 
 from src.configuration import store
-from src.static.static_values_enum import Edition, Element, CardType, Rarity, ManaCap, MatchType
+from src.static.static_values_enum import Edition, Element, CardType, Rarity, ManaCap, MatchType, Format
 
 
 def get_image_url_markdown(card_name, level, edition):
@@ -237,6 +237,28 @@ def filter_mana_cap(input_df, filter_settings):
         return total_df
 
 
+def filter_format(input_df, filter_settings):
+    if input_df.empty:
+        return input_df
+
+    total_df = pd.DataFrame()
+    all_true = True
+    all_false = True
+    for battle_format in Format:
+        active = filter_settings[battle_format.value]
+        if active:
+            all_false = False
+
+            temp_df = input_df.loc[(input_df.format == battle_format.value)]
+            total_df = pd.concat([total_df, temp_df])
+        else:
+            all_true = False
+    if all_true or all_false:
+        return input_df
+    else:
+        return total_df
+
+
 def filter_date(input_df, filter_settings):
     if input_df.empty:
         return input_df
@@ -277,7 +299,7 @@ def get_daily_battle_stats(daily_df):
         daily_df = daily_df.loc[(daily_df.match_type == MatchType.RANKED.value)]
 
         # Select Ranked battles and make dates on day
-        daily_df['created_date'] = pd.to_datetime(daily_df.loc[:, 'created_date']).dt.date
+        daily_df.loc[:, 'created_date'] = pd.to_datetime(daily_df.loc[:, 'created_date']).dt.date
 
         # First group on battle_id
         daily_df = daily_df.groupby(['battle_id'], as_index=False).agg({'result': 'first',