+++ src/static/static_values_enum.py
@@ -24,8 +24,8 @@ class ExtendedEnum(Enum):
 
 
 class Format(ExtendedEnum):
-    MODERN = 'modern'
-    WILD = 'wild'
+    modern = 'modern'
+    wild = 'wild'
 
 
 class MatchType(ExtendedEnum):
@@ -59,12 +59,12 @@ class Leagues(ExtendedEnum):
 
 
 class RatingLevel(ExtendedEnum):
-  Novice = 0
-  Bronze = 1
-  Silver =2
-  Gold = 3
-  Diamond = 4
-  Champion = 5
+    Novice = 0
+    Bronze = 1
+    Silver = 2
+    Gold = 3
+    Diamond = 4
+    Champion = 5
 
 
 class Edition(ExtendedEnum):